EXTCTRL and DRIVER works.
Both EXTCTRL and DRIVER are operational, but require refactoring.
Right now those tasks are just connected with one variable (DRIVER
forwards control input from EXTCTRL to servo motors).

PD0 - pitch signal
PC7 - servo control signal

+++ Inc/task/extctrl.h
@@ -13,9 +13,9 @@
 /* Data struct */
 typedef struct
 {
-    float pitch;
-    float roll;
-    float yaw;
+    uint8_t pitch;
+    uint8_t roll;
+    uint8_t yaw;
 } ExtCtrlData;
 
 /* Data container */
